+++ ft_putnbr_fd.c
@@ -6,12 +6,12 @@
 /*   By: ohaponiu <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/28 15:32:01 by ohaponiu          #+#    #+#             */
-/*   Updated: 2026/03/23 10:13:48 by ohaponiu         ###   ########.fr       */
+/*   Updated: 2026/03/27 11:54:50 by ohaponiu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
-static void	ft_putnbr_fd_rec(unsigned long n, int fd)
+static void	ft_putnbr_fd_rec(unsigned int n, int fd)
 {
 	if (n >= 10)
 		ft_putnbr_fd_rec(n / 10, fd);
@@ -23,8 +23,8 @@ void	ft_putnbr_fd(int n, int fd)
 	if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
-		ft_putnbr_fd_rec(-(unsigned long)n, fd);
+		ft_putnbr_fd_rec(-(unsigned int)n, fd);
 	}
 	else
-		ft_putnbr_fd_rec((unsigned long)n, fd);
+		ft_putnbr_fd_rec((unsigned int)n, fd);
 }
